requirements.txt
seleniumn
selenium-wire
dataclasses
yacs
sqlalchemy
uvicorn